+++ MyLayout/AllTest1ViewController.m
@@ -1,5 +1,5 @@
 //
-//  Test9ViewController.m
+//  AllTest1ViewController.m
 //  MyLayout
 //
 //  Created by oybq on 15/6/21.
@@ -144,8 +144,7 @@
     label2.textColor = [CFTool color:4];
     label2.font = [CFTool font:15];
     label2.myLeftMargin = label2.myRightMargin = 5;
-    label2.numberOfLines = 0;
-    label2.flexedHeight = YES;
+    label2.wrapContentHeight = YES;
     label2.myTopMargin = 10;
     [label2 sizeToFit];
     [tableHeaderViewLayout addSubview:label2];
@@ -199,6 +198,7 @@
 
 -(UITableViewCell*)tableView:(UITableView *)tableView cellForRowAtIndexPath:(NSIndexPath *)indexPath
 {
+    //如果使用了布局来评估cell高度的话，那么请不要使用- (__kindof UITableViewCell *)dequeueReusableCellWithIdentifier:(NSString *)identifier forIndexPath:(NSIndexPath *)indexPath 这个方法，否则可能造成系统崩溃！！！
     AllTest1TableViewCell *cell = (AllTest1TableViewCell*)[tableView dequeueReusableCellWithIdentifier:@"alltest1_cell"];
     if (cell == nil)
         cell = [[AllTest1TableViewCell alloc] initWithStyle:UITableViewCellStyleDefault reuseIdentifier:@"alltest1_cell"];