+++ MyLayout/FOLTest3ViewController.m
@@ -27,9 +27,9 @@
 
 @interface FOLTest3ViewController ()
 
-@property(nonatomic, strong) MyFloatLayout *floatLayout;
+@property(nonatomic, weak) MyFloatLayout *floatLayout;
 
-@property(nonatomic, strong) NSMutableArray *datas;
+@property(nonatomic, strong) NSMutableArray<FOLTest3DataModel*> *datas;
 
 
 @end
@@ -89,12 +89,13 @@
     
     [super viewDidLoad];
     
-    _floatLayout = [MyFloatLayout floatLayoutWithOrientation:MyLayoutViewOrientation_Vert];
-    _floatLayout.myMargin = 0;  //浮动布局和父视图四周的边界是0，也就是说浮动布局的宽度和高度和父视图相等。
-     [self.view addSubview:_floatLayout];
+    MyFloatLayout *floatLayout = [MyFloatLayout floatLayoutWithOrientation:MyLayoutViewOrientation_Vert];
+    floatLayout.myMargin = 0;  //浮动布局和父视图四周的边界是0，也就是说浮动布局的宽度和高度和父视图相等。
+    [self.view addSubview:floatLayout];
+    self.floatLayout = floatLayout;
     
     //通过智能分界线的使用，浮动布局里面的所有子布局视图的分割线都会进行智能的设置，从而解决了我们需求中需要提供边界线的问题。
-    _floatLayout.IntelligentBorderLine = [[MyBorderLineDraw alloc] initWithColor:[[UIColor lightGrayColor] colorWithAlphaComponent:0.2]];
+    floatLayout.IntelligentBorderLine = [[MyBorderLineDraw alloc] initWithColor:[[UIColor lightGrayColor] colorWithAlphaComponent:0.2]];
 
     
     self.navigationItem.rightBarButtonItem = [[UIBarButtonItem alloc] initWithTitle:NSLocalizedString(@"change style", @"") style:UIBarButtonItemStylePlain target:self action:@selector(handleStyleChange:)];
@@ -140,8 +141,7 @@
     titleLabel.text = dataModel.title;
     titleLabel.textColor = [UIColor whiteColor];
     titleLabel.weight = 1;         //向左浮动，宽度和父视图保持一致。
-    titleLabel.numberOfLines = 0;
-    titleLabel.flexedHeight = YES; //如果想让文本消息的高度是动态的，请在设置明确宽度的情况下将numberOfLines设置为0并且将flexedHeight设置为YES。
+    titleLabel.wrapContentHeight = YES; //如果想让文本消息的高度是动态的，请在设置明确宽度的情况下将wrapContentHeight设置为YES。
     [itemLayout addSubview:titleLabel];
     
     itemLayout.gravity = MyMarginGravity_Vert_Bottom; //将整个布局中的所有子视图垂直居底部。
@@ -166,8 +166,7 @@
     titleLabel.text = dataModel.title;
     titleLabel.font = [UIFont boldSystemFontOfSize:17];
     titleLabel.weight =1;   //向左浮动，宽度和父视图保持一致。
-    titleLabel.numberOfLines = 0;
-    titleLabel.flexedHeight = YES; //如果想让文本消息的高度是动态的，请在设置明确宽度的情况下将numberOfLines设置为0并且将flexedHeight设置为YES。
+    titleLabel.wrapContentHeight = YES; //如果想让文本消息的高度是动态的，请在设置明确宽度的情况下将wrapContentHeight设置为YES。
     [itemLayout addSubview:titleLabel];
     
     //来源部分
@@ -202,8 +201,7 @@
     UILabel *titleLabel = [UILabel new];
     titleLabel.text = dataModel.title;
     titleLabel.weight =1;   //向左浮动，宽度和父视图保持一致。
-    titleLabel.numberOfLines = 0;
-    titleLabel.flexedHeight = YES; //如果想让文本消息的高度是动态的，请在设置明确宽度的情况下将numberOfLines设置为0并且将flexedHeight设置为YES。
+    titleLabel.wrapContentHeight = YES; //如果想让文本消息的高度是动态的，请在设置明确宽度的情况下将wrapContentHeight设置为YES。
     [itemLayout addSubview:titleLabel];
     
     //来源部分